--- a/pilot/pkg/model/push_context.go
+++ b/pilot/pkg/model/push_context.go
@@ -37,7 +37,6 @@ import (
 	"istio.io/istio/pkg/config/constants"
 	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/labels"
-	"istio.io/istio/pkg/config/protocol"
 	"istio.io/istio/pkg/config/schema/gvk"
 	"istio.io/istio/pkg/config/schema/kind"
 	"istio.io/istio/pkg/config/visibility"
@@ -1583,38 +1582,29 @@ func SortServicesByCreationTime(services []*Service) []*Service {
 // Caches list of service accounts in the registry
 func (ps *PushContext) initServiceAccounts(env *Environment, services []*Service) {
 	for _, svc := range services {
-		for _, port := range svc.Ports {
-			if port.Protocol == protocol.UDP {
-				continue
-			}
-
-			var accounts sets.String
-			func() {
-				// First get endpoint level service accounts
-				shard, f := env.EndpointIndex.ShardsForService(string(svc.Hostname), svc.Attributes.Namespace)
-				if f {
-					shard.RLock()
-					// copy here to reduce the lock time, the copy cost can be ignored
-					// endpoints could update frequently, so the longer it locks, the more likely it will block other threads.
-					accounts = shard.ServiceAccounts.Copy()
-					shard.RUnlock()
-				}
-				if len(svc.ServiceAccounts) > 0 {
-					if accounts == nil {
-						accounts = sets.New(svc.ServiceAccounts...)
-					} else {
-						accounts = accounts.InsertAll(svc.ServiceAccounts...)
-					}
-				}
-				sa := sets.SortedList(spiffe.ExpandWithTrustDomains(accounts, ps.Mesh.TrustDomainAliases))
-				key := serviceAccountKey{
-					hostname:  svc.Hostname,
-					namespace: svc.Attributes.Namespace,
-				}
-				ps.serviceAccounts[key] = sa
-			}()
-			break
+		var accounts sets.String
+		// First get endpoint level service accounts
+		shard, f := env.EndpointIndex.ShardsForService(string(svc.Hostname), svc.Attributes.Namespace)
+		if f {
+			shard.RLock()
+			// copy here to reduce the lock time
+			// endpoints could update frequently, so the longer it locks, the more likely it will block other threads.
+			accounts = shard.ServiceAccounts.Copy()
+			shard.RUnlock()
 		}
+		if len(svc.ServiceAccounts) > 0 {
+			if accounts == nil {
+				accounts = sets.New(svc.ServiceAccounts...)
+			} else {
+				accounts = accounts.InsertAll(svc.ServiceAccounts...)
+			}
+		}
+		sa := sets.SortedList(spiffe.ExpandWithTrustDomains(accounts, ps.Mesh.TrustDomainAliases))
+		key := serviceAccountKey{
+			hostname:  svc.Hostname,
+			namespace: svc.Attributes.Namespace,
+		}
+		ps.serviceAccounts[key] = sa
 	}
 }