@istio-testing 2023-11-10
@hzxuzhonghu: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
release-notes_istio | 74384fd7309c0b3eb31091e0de42c02618157510 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/47817/release-notes_istio/1722887876963733504) | true | `/test release-notes`
unit-tests-arm64_istio | 74384fd7309c0b3eb31091e0de42c02618157510 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/47817/unit-tests-arm64_istio/1722887863625846784) | true | `/test unit-tests-arm64`
unit-tests_istio | 74384fd7309c0b3eb31091e0de42c02618157510 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/47817/unit-tests_istio/1722887863579709440) | true | `/test unit-tests`
integ-ambient_istio | 74384fd7309c0b3eb31091e0de42c02618157510 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/47817/integ-ambient_istio/1722887872777818112) | true | `/test integ-ambient`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->