@hzxuzhonghu 2023-11-10
with new 

```
cpu: Intel(R) Xeon(R) Gold 6278C CPU @ 2.60GHz
BenchmarkInitServiceAccounts
BenchmarkInitServiceAccounts-4   	  124957	      8981 ns/op	    3840 B/op	      55 allocs/op
PASS
ok  	istio.io/istio/pilot/pkg/model	1.245s
```

Without this patch 

```
cpu: Intel(R) Xeon(R) Gold 6278C CPU @ 2.60GHz
BenchmarkInitServiceAccounts
BenchmarkInitServiceAccounts-4   	  124957	      8981 ns/op	    3840 B/op	      55 allocs/op
PASS
ok  	istio.io/istio/pilot/pkg/model	1.245s
```